Let the user chose a run (graph dataset) to load
Introduced a selected_dataset GET parameter for the graph dashboard. The
value of this parameter is the name of the dataset (run) to load. This
parameter lets users send links to other users of the graph dashboard
that load specific datasets.

When the user changes the dataset in the graph dashboard UI, the URL
parameter correspondingly changes.

If no dataset with the specified name can be found, we show a dialog
that notifies the user. This avoids misleading the user.